+++ core/php/utils.inc.php
@@ -955,7 +955,7 @@ function getNtpTime() {
 				$NTPtime = ord($data[0]) * pow(256, 3) + ord($data[1]) * pow(256, 2) + ord($data[2]) * 256 + ord($data[3]);
 				$TimeFrom1990 = $NTPtime - 2840140800;
 				$TimeNow = $TimeFrom1990 + 631152000;
-				return date("m/d/Y H:i:s", $TimeNow + $time_adjustment);
+				return date("m/d/Y H:i:s", (int) ($TimeNow + $time_adjustment));
 			}
 		}
 	}
